+++ token-variants.mjs
@@ -253,6 +253,50 @@ async function initialize() {
   // Startup ticker that will periodically call 'updateEmbeddedDocuments' with all the accrued updates since the last tick
   startBatchUpdater();
 
+  // Insert default random image field
+  Hooks.on('renderTokenConfig', async (config, html) => {
+    const checkboxRandomize = html.find('input[name="randomImg"]');
+    if (checkboxRandomize.length && !html.find('.token-variants-randomImgDefault').length) {
+      const defaultImg =
+        config.actor?.prototypeToken?.flags['token-variants']?.['randomImgDefault'] ||
+        config.actor?.prototypeToken?.flags['token-hud-wildcard']?.['default'] ||
+        '';
+
+      const field = await renderTemplate(
+        '/modules/token-variants/templates/randomImgDefault.html',
+        { defaultImg, active: checkboxRandomize.is(':checked') }
+      );
+      checkboxRandomize.closest('.form-group').after(field);
+
+      const tvaRngField = html.find('.token-variants-randomImgDefault');
+
+      tvaRngField.find('button').click((event) => {
+        event.preventDefault();
+        const input = tvaRngField.find('input');
+        new FilePicker({ current: input.val(), field: input[0] }).browse(defaultImg);
+      });
+
+      checkboxRandomize.click((event) => {
+        if (event.target.checked) {
+          tvaRngField.addClass('active');
+        } else {
+          tvaRngField.removeClass('active');
+        }
+      });
+    }
+  });
+
+  // Set Default Wildcard images if needed
+  Hooks.on('preCreateToken', (tokenDocument, data, options, userId) => {
+    if (game.user.id === userId && tokenDocument.actor?.prototypeToken?.randomImg) {
+      const defaultImg =
+        tokenDocument.actor?.prototypeToken?.flags['token-variants']?.['randomImgDefault'] ||
+        tokenDocument.actor?.prototypeToken?.flags['token-hud-wildcard']?.['default'] ||
+        '';
+      if (defaultImg) tokenDocument.updateSource({ 'texture.src': defaultImg });
+    }
+  });
+
   Hooks.on('createCombatant', (combatant, options, userId) => {
     if (game.userId !== userId) return;
     const token = combatant._token || canvas.tokens.get(combatant.tokenId);